extract revoke tests to sibling file for anti-monolith ceiling
Move t40-t48 revoke_review_assignment tests from assignment_authority.rs
(2812 LOC) to revoke_assignment_tests.rs sibling module, bringing the
main file to 2500 LOC (ceiling = 2500). Make branch_dir pub(crate) for
sibling test access.


Agend-Agent: archfix-opus-3
Agend-Task: t-20260715012326202220-64650-56
Agend-Branch: fix/review-assignment-revoke-authority
Agend-Issued-At: 2026-07-15T10:17:32.628515+00:00
